fixed bug where swap_dp was using just single precision reals
git-svn-id: svn+ssh://phab.hepforge.org/source/hoppetsvn/branches/qed@284 27345fc9-8513-0410-81e8-f89a8c677c7e

+++ src/sort.f90
@@ -257,8 +257,8 @@ contains
 
 
   subroutine swap_dp(a,b)
-    real, intent(inout) :: a,b
-    real                :: dummy
+    real(dp), intent(inout) :: a,b
+    real(dp)                :: dummy
     dummy=a
     a=b
     b=dummy
